messages/requirements.txt
fastapi[all]
websocket
requests
fastapi-users[sqlalchemy]
pyjwt
passlib[bcrypt]
python-dotenv
psycopg2-binary
pytz